CHROMIUM: rockchip: dmc: Resolve suspend/resume state w/ enable/disable state
There was nothing in the DMC devfreq driver that kept track of whether
we were disabled for suspend/resume or from the enable/disable
notifier.  Since both cases call through to the underlying devfreq code
and the underlying code is not refcounted, that meant that
suspend/resume could confuse the state that was set by the
enable/disable notifier.

Let's just use rockchip_dmc_enable() and rockchip_dmc_disable() in
suspend/resume.  Those are refcounted and so the problem is solved.

This has the side effect of putting us at max speed for suspend/resume,
but that's not necessarily a bad thing other than the small amount of
time spent getting to max speed (a few milliseconds at most).

NOTE: while testing w/ serial console enabled, dmc suspend/resume now
shows as taking a few hundred ms.  That's an artifact of serial
console (which can choose strange places to flush the serial console).
End users don't have serial console enabled, so this won't be a problem
for them.

BUG=chrome-os-partner:45979
TEST=HDMI unplug, suspend, HDMI plug, resume => devfreq disabled & max
TEST=HDMI plug, suspend, HDMI unplug, resume => devfreq enabled
TEST=HDMI unplug, suspend, resume => devfreq enabled
TEST=HDMI plug, suspend, resume => devfreq disabled & max

Change-Id: I757545e4a6dfb2a3c035223b49aa876c8ed02275
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/302808
Reviewed-by: Derek Basehore <dbasehore@chromium.org>