usb: ehci: fix rockchip hsic controller reset bug
In system reboot case, hsic ehci controller has a bug
that reset finished but it's still not enabled (PORT_PE
is zero). If this happens, the ehci driver will set
PORT_OWNER to 1, and the ehci will release ownership of
the port to companion contoller.

But hsic ehci controller actually has no companion contoller,
so it can't release ownership. Generally, the ehci capability
reg HCSPARAMS bit15:12 indicates the number of companion controllers
associated with this USB 2.0 host controller. But the HCSPARAMS reg
of hsic ehci controller has hardware bug, and can't really indicates
the number of companion controllers. so we add a flag manually to
fix the issue.

Signed-off-by: Wu Liang feng <wulf@rock-chips.com>