ARM: dts: garfield: modify the dmc-disable-freq to max
When we plug in a HDMI or other external panel to the board, dmc would
stop to work since that cause screen flick. At this case we should fixed
the ddr frequency to max to get the best performance.

Signed-off-by: Chris Zhong <zyw@rock-chips.com>
(cherry picked from commit 2b4894f41d3888681116266587c3a20094d2bed2)