ARM: dts: rk3288-evb: Add lvds device node on rk3288-evb board
Enable LVDS driver on rk3288-evb board, and attch the "cpt,claa070wp03xg"
simple panel to it.

Note that lvds-panel and edp-panel share the same enable-gpios, and one
gpio couldn't request for twice, so just a hack that I delete the enable-gpios
numbers from lvds-panel.

Due to there are only two vop module, that's to say we can't keep enable
eDP / LVDS / HDMI at the same time, so this time we still keep LVDS device
disabled. So if you want to test lvds device, then you should disable the
HDMI or eDP device, and enable the LVDS device, like:

 &hdmi {
        ddc-i2c-bus = <&i2c5>;
-       status = "okay";
+       status = "disabled";
 };

 &lvds {
        rockchip,data-width = <24>;
        rockchip,output = "lvds";
        rockchip,panel = <&lvds_panel>;
-       status = "disabled";
+       status = "okay";
 };

Signed-off-by: Yakir Yang <ykk@rock-chips.com>