FROMLIST: ARM: dts: rockchip: override edp hpd handling on veyron-pinky and speedy
Pinky boards don't have the hotplug pin connected. So remove the
hotplug pinctrl setting and enable the force-hpd option, to allow
them to find the display too.

While on speedy boards, the hotplug pin is connected, judging by comments
in a chromeos change it seems the "panels HPD voltage is too low to be
detected", so it also needs the forced hotplug, as we of course also know
that a display is connected.

Change-Id: Ia442b505fd08855a20530cf1b9fe4e54f6bdb96e

Signed-off-by: Heiko Stuebner <heiko@sntech.de>

(am from https://patchwork.kernel.org/patch/8615571/)
Signed-off-by: Caesar Wang <wxt@rock-chips.com>